+++ src/components/Visualizations/Graph/HistogramGraph.jsx
@@ -1,41 +1,7 @@
 import React from 'react';
 import { Graph } from 'components/Visualizations';
 import strings from 'lang';
-
-/**
- * Converts an HSV color value to RGB. Conversion formula
- * adapted from http://en.wikipedia.org/wiki/HSV_color_space.
- * Assumes h, s, and v are contained in the set [0, 1] and
- * returns r, g, and b in the set [0, 255].
- *
- * @param   Number  h       The hue
- * @param   Number  s       The saturation
- * @param   Number  v       The value
- * @return  Array           The RGB representation
- */
-function hsvToRgb(h, s, v) {
-  let r;
-  let g;
-  let b;
-
-  const i = Math.floor(h * 6);
-  const f = (h * 6) - i;
-  const p = v * (1 - s);
-  const q = v * (1 - (f * s));
-  const t = v * (1 - ((1 - f) * s));
-
-  switch (i % 6) {
-    case 0: r = v; g = t; b = p; break;
-    case 1: r = q; g = v; b = p; break;
-    case 2: r = p; g = v; b = t; break;
-    case 3: r = p; g = q; b = v; break;
-    case 4: r = t; g = p; b = v; break;
-    case 5: r = v; g = p; b = q; break;
-    default: r = v; g = t; b = p;
-  }
-
-  return [r * 255, g * 255, b * 255];
-}
+import { hsvToRgb } from 'utility';
 
 const colorFn = columns => (color, data) => {
   if (data.index || data.index === 0) {
